Move traffic_monitor log to generic location
