+++ traffic_monitor_golang/traffic_monitor/health/event.go
@@ -24,7 +24,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/log"
+	"github.com/apache/incubator-trafficcontrol/lib/go-log"
 )
 
 type Time time.Time
