+++ traffic_monitor_golang/traffic_monitor/srvhttp/srvhttp.go
@@ -28,7 +28,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/log"
+	"github.com/apache/incubator-trafficcontrol/lib/go-log"
 	"github.com/hydrogen18/stoppableListener"
 )
