+++ traffic_ops/traffic_ops_golang/auth.go
@@ -22,7 +22,8 @@ package main
 import (
 	"database/sql"
 
-	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/log"
+	"github.com/apache/incubator-trafficcontrol/lib/go-log"
+
 	"github.com/jmoiron/sqlx"
 )
