+++ traffic_ops/traffic_ops_golang/cdns.go
@@ -25,8 +25,9 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/log"
+	"github.com/apache/incubator-trafficcontrol/lib/go-log"
 	"github.com/apache/incubator-trafficcontrol/traffic_ops/tostructs"
+
 	"github.com/jmoiron/sqlx"
 )
