+++ traffic_ops/traffic_ops_golang/monitoring.go
@@ -30,7 +30,7 @@ import (
 	"github.com/jmoiron/sqlx"
 	"github.com/lib/pq"
 
-	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/log"
+	"github.com/apache/incubator-trafficcontrol/lib/go-log"
 )
 
 const MonitoringPrivLevel = PrivLevelReadOnly
