+++ traffic_ops/traffic_ops_golang/routes.go
@@ -28,7 +28,7 @@ import (
 	"net/http/httputil"
 	"time"
 
-	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/log"
+	"github.com/apache/incubator-trafficcontrol/lib/go-log"
 )
 
 // Routes returns the routes, and a catchall route for when no route matches.
@@ -67,11 +67,11 @@ func rootHandler(d ServerData) http.Handler {
 	// debug
 	tr := &http.Transport{
 		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-		DialContext:     (&net.Dialer{
-			Timeout: time.Duration(d.Config.ProxyTimeout) * time.Second,
+		DialContext: (&net.Dialer{
+			Timeout:   time.Duration(d.Config.ProxyTimeout) * time.Second,
 			KeepAlive: time.Duration(d.Config.ProxyKeepAlive) * time.Second,
 		}).DialContext,
-		TLSHandshakeTimeout: time.Duration(d.Config.ProxyTLSTimeout) * time.Second,
+		TLSHandshakeTimeout:   time.Duration(d.Config.ProxyTLSTimeout) * time.Second,
 		ResponseHeaderTimeout: time.Duration(d.Config.ProxyReadHeaderTimeout) * time.Second,
 		//Other knobs we can turn: ExpectContinueTimeout,IdleConnTimeout
 	}
@@ -79,7 +79,7 @@ func rootHandler(d ServerData) http.Handler {
 	rp.Transport = tr
 
 	rp.ErrorLog = log.Error //if we don't provide a logger to the reverse proxy it logs to stdout/err and is lost when ran by a script.
-	log.Debugf("our reverseProxy: %++v\n",rp)
+	log.Debugf("our reverseProxy: %++v\n", rp)
 	log.Debugf("our reverseProxy's transport: %++v\n", tr)
 	loggingProxyHandler := wrapAccessLog(d.TOSecret, rp)
 	return loggingProxyHandler
