+++ traffic_ops/traffic_ops_golang/routing.go
@@ -26,7 +26,8 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/log"
+	"github.com/apache/incubator-trafficcontrol/lib/go-log"
+
 	"github.com/jmoiron/sqlx"
 )
