+++ traffic_ops/traffic_ops_golang/wrappers.go
@@ -31,7 +31,7 @@ import (
 	"time"
 	"unicode"
 
-	"github.com/apache/incubator-trafficcontrol/traffic_monitor_golang/common/log"
+	"github.com/apache/incubator-trafficcontrol/lib/go-log"
 	"github.com/apache/incubator-trafficcontrol/traffic_ops/tocookie"
 )
